requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-core
faiss-cpu
pypdf
tiktoken
python-dotenv